requirements.txt
fastapi
uvicorn
python-multipart
easyocr
openai
Pillow
numpy
python-dotenv